feat: rename ai terminal summary extension to ai terminal assistant extension
