+++ examples/browser/tsconfig.json
@@ -78,7 +78,7 @@
       "path": "../../packages/ai-terminal"
     },
     {
-      "path": "../../packages/ai-terminal-summary"
+      "path": "../../packages/ai-terminal-assistant"
     },
     {
       "path": "../../packages/ai-terminal-summary-addons"
