+++ examples/electron/tsconfig.json
@@ -78,9 +78,6 @@
       "path": "../../packages/ai-terminal"
     },
     {
-      "path": "../../packages/ai-terminal-summary"
-    },
-    {
       "path": "../../packages/ai-terminal-summary-addons"
     },
     {
