+++ packages/ai-terminal-assistant/.eslintrc.js
