+++ packages/ai-terminal-assistant/src/browser/ai-terminal-assistant-command-service.ts
