+++ packages/ai-terminal-assistant/src/browser/ai-terminal-assistant-contribution.ts
@@ -16,16 +16,16 @@
 
 import { inject, injectable } from '@theia/core/shared/inversify';
 import { TerminalService } from '@theia/terminal/lib/browser/base/terminal-service';
-import { AiTerminalSummaryAgent } from './ai-terminal-summary-agent';
+import { AiTerminalSummaryAgent } from './terminal-output-analysis-agent';
 import { ENABLE_AI_CONTEXT_KEY } from '@theia/ai-core/lib/browser';
 import { AICommandHandlerFactory } from '@theia/ai-core/lib/browser/ai-command-handler-factory';
 import { AgentService } from '@theia/ai-core';
 import { ApplicationShell, codicon, KeybindingContribution, KeybindingRegistry, WidgetManager } from '@theia/core/lib/browser';
 import { Command, CommandContribution, CommandRegistry, MenuContribution, MenuModelRegistry, nls } from '@theia/core';
 import { TerminalMenus } from '@theia/terminal/lib/browser/terminal-frontend-contribution';
-import { SUMMARY_VIEW_WIDGET_ID } from '../common/summary-view-widget';
-import { SummaryViewWidget } from './summary-view-widget';
-import { SummaryService } from './summary-service';
+import { SUMMARY_VIEW_WIDGET_ID } from '../common/ai-terminal-assistant-view-widget';
+import { SummaryViewWidget } from './ai-terminal-assistant-view-widget';
+import { SummaryService } from './ai-terminal-assistant-service';
 
 export const AI_TERMINAL_SUMMARY_TOGGLE_COMMAND_ID = 'aiTerminalSummary:toggle';
