+++ packages/ai-terminal-assistant/src/browser/style/ai-terminal-assistant.css
