+++ packages/ai-terminal-assistant/src/browser/style/monaco-decorations.css
