+++ packages/ai-terminal-assistant/src/browser/terminal-output-analysis-agent.ts
@@ -25,7 +25,7 @@ import {
 } from '@theia/ai-core/lib/common';
 import { LanguageModelService } from '@theia/ai-core/lib/browser';
 import { generateUuid, ILogger, nls } from '@theia/core';
-import { terminalPrompts } from './ai-terminal-summary-prompt-template';
+import { terminalPrompts } from './ai-terminal-assistant-prompt-template';
 import { inject, injectable } from '@theia/core/shared/inversify';
 import { z } from 'zod';
 import zodToJsonSchema from 'zod-to-json-schema';
