+++ packages/ai-terminal-assistant/src/common/ai-terminal-assistant-view-widget.ts
