+++ packages/ai-terminal-assistant/tsconfig.json
