+++ packages/ai-terminal-summary-addons/package.json
@@ -5,7 +5,7 @@
   "dependencies": {
     "@theia/ai-chat": "1.70.0",
     "@theia/ai-core": "1.70.0",
-    "@theia/ai-terminal-summary": "1.70.0",
+    "@theia/ai-terminal-assistant": "1.70.0",
     "@theia/core": "1.70.0",
     "@theia/terminal": "1.70.0",
     "@theia/debug": "1.70.0"
