+++ packages/ai-terminal-summary-addons/tsconfig.json
@@ -16,7 +16,7 @@
       "path": "../ai-core"
     },
     {
-      "path": "../ai-terminal-summary"
+      "path": "../ai-terminal-assistant"
     },
     {
       "path": "../core"
